+++ dalesbred/src/test/java/fi/evident/dalesbred/TransactionSettingsTest.java
@@ -22,9 +22,13 @@
 
 package fi.evident.dalesbred;
 
+import org.jetbrains.annotations.NotNull;
 import org.junit.Test;
 
+import java.lang.annotation.Annotation;
+
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertSame;
 
 public class TransactionSettingsTest {
 
@@ -42,5 +46,47 @@ public class TransactionSettingsTest {
     @Test(expected=IllegalArgumentException.class)
     public void negativeRetries() {
         new TransactionSettings().setRetries(-1);
+    }
+
+    @Test
+    public void initializeFromAnnotation() {
+        MyTransactional transactional = new MyTransactional();
+        transactional.isolation = Isolation.SERIALIZABLE;
+        transactional.propagation = Propagation.NESTED;
+        transactional.retries = 4;
+
+        TransactionSettings settings = TransactionSettings.fromAnnotation(transactional);
+        assertSame(transactional.isolation(), settings.getIsolation());
+        assertSame(transactional.propagation(), settings.getPropagation());
+        assertSame(transactional.retries(), settings.getRetries());
+    }
+
+    @SuppressWarnings("ClassExplicitlyAnnotation")
+    private static final class MyTransactional implements Transactional {
+        Isolation isolation = Isolation.REPEATABLE_READ;
+        Propagation propagation = Propagation.REQUIRED;
+        int retries = 0;
+
+        @NotNull
+        @Override
+        public Propagation propagation() {
+            return propagation;
+        }
+
+        @NotNull
+        @Override
+        public Isolation isolation() {
+            return isolation;
+        }
+
+        @Override
+        public int retries() {
+            return retries;
+        }
+
+        @Override
+        public Class<? extends Annotation> annotationType() {
+            return Transactional.class;
+        }
     }
 }